Don't include system structs into our own classes
Previously, we included this struct into the CLog class. That triggered
a Coverity warning for uninitialized members.

In and of itself, it's not really a problem. However, since struct flock
is really only ever passed on as a pointer anyway, it seems to make more
sense to declare it in our class as a pointer, rather than a struct. The
result is also that our class size isn't dependent on
implementation-defined struct sizes.

Since it's a pointer, we can now also initialize it to NULL, thereby
silencing the Coverity warning

Coverity CID#18510