Remove compiler warnings
A few final places of string literals being assigned to char* (rather
than const char*) variables.

Unfortunately we don't really fix the problem, since we pass it on to a
function that does not have the const attribute set for the relevant
parameters, and therefore we need to cast that attribute away.

At least we're reasonably sure it doesn't happen in practice, though.